Add full functionality for active learning.
This is a (probably buggy) first attempt. It adds the example with the
highest entropy to the training pool after one epoch of training. Should
possibly/probably use multiple epochs of training before adding new
examples, but it serves as a PoC. Testing hopefully will show that this
outperforms random sampling.