Refactor active selection code.
Selectors are now objects instead of chains of ifs. This is done to
facilitate easy adding of new active methods for testing. The selectors
get injected with the trainer object, so they have full access to all
variables. They are called through the `select()` function, that returns
an index for which unlabeled example to copy. In the future, multiple
indices should probably be returned, in the case where one wants to copy
more than one example at a time.